requirements.txt
discord.py
torch
transformers